Move response channel to util mod for now (since it's used with streaming and non streaming)
