+++ mistralrs-server-core/src/streaming.rs
@@ -2,16 +2,6 @@
 
 use std::env;
 
-use mistralrs_core::Response;
-use tokio::sync::mpsc::{channel, Receiver, Sender};
-
-/// Default buffer size for the response channel used in streaming operations.
-///
-/// This constant defines the maximum number of response messages that can be buffered
-/// in the channel before backpressure is applied. A larger buffer reduces the likelihood
-/// of blocking but uses more memory.
-pub const DEFAULT_CHANNEL_BUFFER_SIZE: usize = 10_000;
-
 /// Default keep-alive interval for Server-Sent Events (SSE) streams in milliseconds.
 pub const DEFAULT_KEEP_ALIVE_INTERVAL_MS: u64 = 10_000;
 
@@ -23,15 +13,6 @@ pub enum DoneState {
     SendingDone,
     /// The stream has completed entirely
     Done,
-}
-
-/// Creates a channel for response communication.
-pub fn create_response_channel(
-    buffer_size: Option<usize>,
-) -> (Sender<Response>, Receiver<Response>) {
-    let channel_buffer_size = buffer_size.unwrap_or(DEFAULT_CHANNEL_BUFFER_SIZE);
-
-    channel(channel_buffer_size)
 }
 
 /// Gets the keep-alive interval for SSE streams from environment or default.
